shooting in front of tank with different angles works but shooting back still has works

+++ src/game/elements/Tank.java
@@ -17,7 +17,7 @@ public class Tank extends GameObject{
     private int type;
 
     //constructor
-    public Tank(int x, int y, int health, int type){ // gun must be added manually after making tank
+    public Tank(double x, double y, int health, int type){ // gun must be added manually after making tank
         super(x,y);
         this.health = health;
         this.type = type;
